Did TODO task

+++ src/DataTypes.java
@@ -4,7 +4,7 @@ public class DataTypes {
     // TASK 1: fix this code so that it passes the test in DataTypesTest.java
     public static long sum(List<Integer> numbers) {
 
-        long s = 0;
+        int s = 0;
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
